Modified the app to allow 2 different instances
Quick hack to allow users with 2 pedals to run a version of the original program as well as this version (one for each pedal).

Modified the app to grab the last pedal vs the first pedal found then updated the names everywhere so it won't be seen as a duplicate instance.

Had to change a few other things to get it to compile with latest .Net as well.